extracted automerge init from handler init, fixed target keys
