+++ packages/core/echo/echo-schema/src/echo-handler/create.ts
@@ -6,7 +6,7 @@ import type * as S from '@effect/schema/Schema';
 
 import { ComplexMap } from '@dxos/util';
 
-import { DATA_NAMESPACE, EchoReactiveHandler } from './echo-handler';
+import { DATA_NAMESPACE, EchoReactiveHandler, PROPERTY_ID, throwIfCustomClass } from './echo-handler';
 import {
   type ObjectInternals,
   type ProxyTarget,
@@ -23,6 +23,7 @@ import {
   createReactiveProxy,
   getProxyHandlerSlot,
   isReactiveObject,
+  DynamicEchoSchema,
 } from '../ddl';
 import type { EchoReactiveObject, ObjectMeta } from '../ddl';
 
@@ -34,6 +35,7 @@ export const createEchoObject = <T extends {}>(init: T): EchoReactiveObject<T> =
   if (schema != null) {
     validateSchema(schema);
   }
+  validateInitialProps(init);
 
   if (isReactiveObject(init)) {
     const proxy = init as any;
@@ -54,7 +56,10 @@ export const createEchoObject = <T extends {}>(init: T): EchoReactiveObject<T> =
     target[symbolPath] = [];
     target[symbolNamespace] = DATA_NAMESPACE;
     slot.handler._proxyMap.set(target, proxy);
-    slot.handler._init(target);
+
+    initCore(core, target);
+    slot.handler.init(target);
+
     saveTypeInAutomerge(target[symbolInternals], schema);
     if (meta && meta.keys.length > 0) {
       target[symbolInternals].core.setMeta(meta);
@@ -71,11 +76,21 @@ export const createEchoObject = <T extends {}>(init: T): EchoReactiveObject<T> =
       [symbolNamespace]: DATA_NAMESPACE,
       ...(init as any),
     };
+    initCore(core, target);
     const proxy = createReactiveProxy<ProxyTarget>(target, EchoReactiveHandler.instance) as any;
     core.rootProxy = proxy;
     saveTypeInAutomerge(target[symbolInternals], schema);
     return proxy;
   }
+};
+
+const initCore = (core: AutomergeObjectCore, target: any) => {
+  // Handle ID pre-generated by `create`.
+  if (PROPERTY_ID in target) {
+    target[symbolInternals].core.id = target[PROPERTY_ID];
+    delete target[PROPERTY_ID];
+  }
+  core.initNewObject(target);
 };
 
 export const initEchoReactiveObjectRootProxy = (core: AutomergeObjectCore) => {
@@ -98,5 +113,21 @@ const validateSchema = (schema: S.Schema<any>) => {
 const saveTypeInAutomerge = (internals: ObjectInternals, schema: S.Schema<any> | undefined) => {
   if (schema != null) {
     internals.core.setType(requireTypeReference(schema));
+  }
+};
+
+const validateInitialProps = (target: any) => {
+  for (const key in target) {
+    const value = target[key];
+    if (value === undefined) {
+      delete target[key];
+    } else if (typeof value === 'object') {
+      if (value instanceof DynamicEchoSchema) {
+        target[key] = value.serializedSchema;
+      } else {
+        throwIfCustomClass(key, value);
+      }
+      validateInitialProps(target[key]);
+    }
   }
 };
