+++ packages/core/echo/echo-schema/src/echo-handler/echo-handler.ts
@@ -3,7 +3,6 @@
 //
 
 import * as S from '@effect/schema/Schema';
-import { type Brand } from 'effect';
 import { inspect, type InspectOptionsStylized } from 'node:util';
 
 import { Reference } from '@dxos/echo-db';
@@ -19,6 +18,7 @@ import {
   symbolPath,
   symbolHandler,
   type ObjectInternals,
+  TargetKey,
 } from './echo-proxy-target';
 import { type AutomergeObjectCore, META_NAMESPACE } from '../automerge/automerge-object-core';
 import { type KeyPath } from '../automerge/key-path';
@@ -26,34 +26,7 @@ import { SchemaValidator, DynamicEchoSchema, StoredEchoSchema } from '../ddl';
 import { createReactiveProxy, symbolIsProxy, type ReactiveHandler, type ObjectMeta } from '../ddl';
 import { defineHiddenProperty } from '../util/property';
 
-/**
- * Extends the native array with methods overrides for automerge.
- */
-// TODO(dmaretskyi): Rename once the original AutomergeArray gets deleted.
-
-/**
- * For tracking proxy targets in the `targetsMap`.
- */
-type TargetKey = {
-  path: KeyPath;
-  namespace: string;
-  type: 'record' | 'array';
-} & Brand.Brand<'TargetKey'>;
-
-const TargetKey = {
-  /**
-   * Constructor function forces the order of the fields.
-   */
-  new: (path: KeyPath, namespace: string, type: 'record' | 'array'): TargetKey =>
-    ({
-      path,
-      namespace,
-      type,
-    }) as TargetKey,
-  hash: (key: TargetKey): string => JSON.stringify(key),
-};
-
-const PROPERTY_ID = 'id';
+export const PROPERTY_ID = 'id';
 
 export const DATA_NAMESPACE = 'data';
 
@@ -65,23 +38,10 @@ export class EchoReactiveHandler implements ReactiveHandler<ProxyTarget> {
 
   _proxyMap = new WeakMap<object, any>();
 
-  _init(target: ProxyTarget): void {
+  init(target: ProxyTarget): void {
     invariant(target[symbolInternals]);
     invariant(!(target as any)[symbolIsProxy]);
     invariant(Array.isArray(target[symbolPath]));
-
-    // Handle ID pre-generated by `create`.
-    if (PROPERTY_ID in target) {
-      target[symbolInternals].core.id = target[PROPERTY_ID];
-      delete target[PROPERTY_ID];
-    }
-
-    if (target[symbolPath].length === 0 && target[symbolNamespace] === DATA_NAMESPACE) {
-      this.validateInitialProps(target);
-      if (target[symbolInternals].core.database == null) {
-        target[symbolInternals].core.initNewObject(target);
-      }
-    }
 
     // Clear extra keys from objects
     if (!Array.isArray(target)) {
@@ -95,22 +55,6 @@ export class EchoReactiveHandler implements ReactiveHandler<ProxyTarget> {
     defineHiddenProperty(target, symbolHandler, this);
     if (inspect.custom) {
       defineHiddenProperty(target, inspect.custom, this._inspect.bind(target));
-    }
-  }
-
-  private validateInitialProps(target: ProxyTarget) {
-    for (const key in target) {
-      const value = target[key];
-      if (value === undefined) {
-        delete target[key];
-      } else if (typeof value === 'object') {
-        if (value instanceof DynamicEchoSchema) {
-          target[key] = value.serializedSchema;
-        } else {
-          throwIfCustomClass(key, value);
-        }
-        this.validateInitialProps(target[key]);
-      }
     }
   }
 
@@ -182,7 +126,8 @@ export class EchoReactiveHandler implements ReactiveHandler<ProxyTarget> {
       return this._handleStoredSchema(target, target[symbolInternals].core.lookupLink(decoded));
     }
     if (Array.isArray(decoded)) {
-      const newTarget = defaultMap(target[symbolInternals].targetsMap, dataPath, (): ProxyTarget => {
+      const targetKey = TargetKey.new(dataPath, namespace, 'array');
+      const newTarget = defaultMap(target[symbolInternals].targetsMap, targetKey, (): ProxyTarget => {
         const array = new EchoArray();
         array[symbolInternals] = target[symbolInternals];
         array[symbolPath] = dataPath;
@@ -193,10 +138,11 @@ export class EchoReactiveHandler implements ReactiveHandler<ProxyTarget> {
       return createReactiveProxy(newTarget, this);
     }
     if (typeof decoded === 'object') {
+      const targetKey = TargetKey.new(dataPath, namespace, 'record');
       // TODO(dmaretskyi): Materialize properties for easier debugging.
       const newTarget = defaultMap(
         target[symbolInternals].targetsMap,
-        dataPath,
+        targetKey,
         (): ProxyTarget => ({
           [symbolInternals]: target[symbolInternals],
           [symbolPath]: dataPath,
@@ -555,7 +501,7 @@ export class EchoReactiveHandler implements ReactiveHandler<ProxyTarget> {
   }
 }
 
-const throwIfCustomClass = (prop: KeyPath[number], value: any) => {
+export const throwIfCustomClass = (prop: KeyPath[number], value: any) => {
   if (value == null || Array.isArray(value)) {
     return;
   }
